Move wallet Lightning Address lookups into CRUD
