+++ lnbits/core/crud/wallets.py
@@ -1,5 +1,6 @@
 from datetime import datetime, timezone
 from time import time
+from typing import Any
 from uuid import uuid4
 
 from lnbits.core.db import db
@@ -227,6 +228,50 @@ async def get_wallets_count():
     result = await db.execute("SELECT COUNT(*) as count FROM wallets")
     row = result.mappings().first()
     return row.get("count", 0)
+
+
+async def wallet_lightning_address_exists(
+    local_part: str, conn: Connection | None = None
+) -> bool:
+    row: Any = await (conn or db).fetchone(
+        """
+        SELECT id FROM wallets
+        WHERE lightning_address = :lightning_address
+        """,
+        {"lightning_address": local_part},
+    )
+    return bool(row)
+
+
+async def wallet_lightning_address_exists_for_other_wallet(
+    local_part: str, wallet_id: str, conn: Connection | None = None
+) -> bool:
+    row: Any = await (conn or db).fetchone(
+        """
+        SELECT id FROM wallets
+        WHERE lightning_address = :lightning_address
+            AND id != :wallet_id
+        """,
+        {"lightning_address": local_part, "wallet_id": wallet_id},
+    )
+    return bool(row)
+
+
+async def get_wallet_by_lightning_address(local_part: str) -> Wallet | None:
+    return await db.fetchone(
+        """
+        SELECT wallets.*, COALESCE((
+            SELECT balance FROM balances WHERE wallet_id = wallets.id
+        ), 0) AS balance_msat FROM wallets
+        INNER JOIN accounts ON wallets.user = accounts.id
+        WHERE lightning_address = :lightning_address
+            AND wallet_type = 'lightning'
+            AND deleted = false
+            AND accounts.activated = true
+        """,
+        {"lightning_address": local_part.lower()},
+        Wallet,
+    )
 
 
 async def get_wallet_for_key(
